+++ docs/usage/static-files.rst
@@ -86,8 +86,8 @@ under which a specific file will be available:
 ---------------------
 
 To customize how Litestar interacts with the file system, a class implementing the
-:class:`~litestar.types.BaseFileSystem` can be passed to ``file_system``. An example
-of this are the file systems provided by
+:class:`~litestar.file_system.BaseFileSystem` can be passed to ``file_system``. An
+example of this are the file systems provided by
 `fsspec <https://filesystem-spec.readthedocs.io/en/latest/>`_, which includes support
 for FTP, SFTP, Hadoop, SMB, GitHub and
 `many more <https://filesystem-spec.readthedocs.io/en/latest/api.html#implementations>`_,
