+++ litestar/types/__init__.py
@@ -74,7 +74,6 @@ from .composite_types import (
 )
 from .debugger_types import Debugger
 from .empty import Empty, EmptyType
-from .file_types import BaseFileSystem, FileInfo
 from .helper_types import AnyIOBackend, MaybePartial, OptionalSequence, SSEData, StreamType, SyncOrAsyncUnion
 from .internal_types import ControllerRouterHandler, ReservedKwargs, RouteHandlerMapItem, RouteHandlerType
 from .protocols import DataclassProtocol, Logger
@@ -90,7 +89,6 @@ __all__ = (
     "AnyGenerator",
     "AnyIOBackend",
     "AsyncAnyCallable",
-    "BaseFileSystem",
     "BaseScope",
     "BeforeMessageSendHookHandler",
     "BeforeRequestHookHandler",
@@ -104,7 +102,6 @@ __all__ = (
     "EmptyType",
     "ExceptionHandler",
     "ExceptionHandlersMap",
-    "FileInfo",
     "GetLogger",
     "Guard",
     "HTTPDisconnectEvent",
