+++ tests/unit/test_static_files/conftest.py
@@ -4,8 +4,7 @@ import pytest
 from _pytest.fixtures import FixtureRequest
 from fsspec.implementations.local import LocalFileSystem
 
-from litestar.file_system import BaseLocalFileSystem, maybe_wrap_fsspec_file_system
-from litestar.types import BaseFileSystem
+from litestar.file_system import BaseFileSystem, BaseLocalFileSystem, maybe_wrap_fsspec_file_system
 
 
 @pytest.fixture(params=[BaseLocalFileSystem(), LocalFileSystem()])
